+++ rotkehlchen/tests/unit/test_cost_basis.py
@@ -5,6 +5,7 @@ from pathlib import Path
 
 import pytest
 
+from rotkehlchen.accounting.accountant import Accountant
 from rotkehlchen.accounting.cost_basis import AssetAcquisitionEvent
 from rotkehlchen.accounting.export.csv import FILENAME_ALL_CSV
 from rotkehlchen.accounting.mixins.event import AccountingEventType
@@ -18,9 +19,20 @@ from rotkehlchen.chain.evm.accounting.structures import TxAccountingTreatment, T
 from rotkehlchen.constants.assets import A_3CRV, A_BTC, A_ETH, A_EUR, A_WETH
 from rotkehlchen.constants.misc import ONE, ZERO
 from rotkehlchen.db.settings import DBSettings
+from rotkehlchen.exchanges.data_structures import Trade
 from rotkehlchen.fval import FVal
+from rotkehlchen.tests.utils.accounting import accounting_history_process
 from rotkehlchen.tests.utils.factories import make_evm_address, make_random_bytes
-from rotkehlchen.types import CostBasisMethod, Location, Timestamp, make_evm_tx_hash
+from rotkehlchen.types import (
+    AssetAmount,
+    CostBasisMethod,
+    Fee,
+    Location,
+    Price,
+    Timestamp,
+    TradeType,
+    make_evm_tx_hash,
+)
 
 EXAMPLE_TIMESTAMP = Timestamp(1675483017)
 
@@ -1047,3 +1059,86 @@ def test_taxable_acquisition(accountant):
     assert pot.processed_events[0].free_amount == ZERO
     assert pot.processed_events[0].pnl.taxable == ONE
     assert pot.processed_events[0].pnl.free == ZERO
+
+
+@pytest.mark.parametrize('mocked_price_queries', [{
+    A_ETH: {
+        A_EUR: {
+            1677593073: FVal(50),
+            1677593074: FVal(120),
+            1677593075: FVal(130),
+            1677593076: FVal(90),
+        },
+    },
+}])
+@pytest.mark.parametrize(('db_settings', 'expected_pnls'), [
+    (
+        {'cost_basis_method': CostBasisMethod.FIFO, 'include_fees_in_cost_basis': False},
+        [ZERO, ZERO, FVal(-10), FVal(3500), ZERO, FVal(-10), ZERO, ZERO, FVal(-10), FVal(1600), ZERO, FVal(-10)],  # noqa: E501
+    ),
+    (
+        {'cost_basis_method': CostBasisMethod.ACB, 'include_fees_in_cost_basis': True},
+        [ZERO, ZERO, ZERO, FVal(3485), ZERO, ZERO, ZERO, ZERO, ZERO, FVal(-16), ZERO, ZERO],
+    ),
+])
+def test_fees(accountant: 'Accountant', expected_pnls: list[FVal]):
+    """
+    Tests that fees are properly either calculated as standalone events or included in the price.
+    Values for the example are taken from the Canada example from this issue comment
+    https://github.com/rotki/rotki/issues/5561#issuecomment-1423338938.
+    """
+    history = [
+        Trade(
+            timestamp=Timestamp(1677593073),
+            location=Location.EXTERNAL,
+            base_asset=A_ETH,
+            quote_asset=A_EUR,
+            trade_type=TradeType.BUY,
+            amount=AssetAmount(FVal(100)),
+            rate=Price(FVal(50)),
+            fee=Fee(FVal(10)),
+            fee_currency=A_EUR,
+            notes='Trade 1',
+        ), Trade(
+            timestamp=Timestamp(1677593074),
+            location=Location.EXTERNAL,
+            base_asset=A_ETH,
+            quote_asset=A_EUR,
+            trade_type=TradeType.SELL,
+            amount=AssetAmount(FVal(50)),
+            rate=Price(FVal(120)),
+            fee=Fee(FVal(10)),
+            fee_currency=A_EUR,
+            notes='Trade 2',
+        ), Trade(
+            timestamp=Timestamp(1677593075),
+            location=Location.EXTERNAL,
+            base_asset=A_ETH,
+            quote_asset=A_EUR,
+            trade_type=TradeType.BUY,
+            amount=AssetAmount(FVal(50)),
+            rate=Price(FVal(130)),
+            fee=Fee(FVal(10)),
+            fee_currency=A_EUR,
+            notes='Trade 3',
+        ), Trade(
+            timestamp=Timestamp(1677593076),
+            location=Location.EXTERNAL,
+            base_asset=A_ETH,
+            quote_asset=A_EUR,
+            trade_type=TradeType.SELL,
+            amount=AssetAmount(FVal(40)),
+            rate=Price(FVal(90)),
+            fee=Fee(FVal(10)),
+            fee_currency=A_EUR,
+            notes='Trade 4',
+        ),
+    ]
+    accounting_history_process(
+        accountant=accountant,
+        start_ts=Timestamp(0),
+        end_ts=Timestamp(1677593077),
+        history_list=history,  # type: ignore[arg-type]  # invariant problem
+    )
+    for event, expected_pnl in zip(accountant.pots[0].processed_events, expected_pnls):
+        assert event.pnl.taxable == expected_pnl